requirements.txt
python-gitlab==1.6.0
# python-ldap==3.4.0
ldap3
